MOBILE-4749 rte: Refactor to allow alternative implementations
