+++ src/core/features/editor/components/rich-text-editor/core-rich-text-editor.html
@@ -1,0 +1,2 @@
+<core-dynamic-component [component]="editorComponentClass" [data]="editorComponentData" #dynamicComponent />
+<input type="hidden" [name]="name" [formControl]="control" />
