+++ src/core/features/editor/components/rich-text-editor/rich-text-editor.ts
@@ -13,55 +13,52 @@
 // limitations under the License.
 
 import {
+    AfterViewInit,
     Component,
+    CUSTOM_ELEMENTS_SCHEMA,
+    ElementRef,
+    OnDestroy,
+    Optional,
+    Type,
+    ViewChild, OnInit,
     Input,
     Output,
     EventEmitter,
-    ViewChild,
-    ElementRef,
-    OnInit,
-    OnDestroy,
-    Optional,
-    AfterViewInit,
-    CUSTOM_ELEMENTS_SCHEMA,
 } from '@angular/core';
+import { IonContent } from '@ionic/angular';
+import { CoreSharedModule } from '@/core/shared.module';
+import { CoreDynamicComponent } from '@components/dynamic-component/dynamic-component';
+import { CoreEditorBaseComponent } from '@features/editor/classes/base-editor-component';
+import { CoreEventFormActionData, CoreEventObserver, CoreEvents } from '@singletons/events';
+import { CoreWait } from '@singletons/wait';
+import { CoreDom } from '@singletons/dom';
+import { ContextLevel } from '@/core/constants';
 import { FormControl } from '@angular/forms';
-import { IonTextarea, IonContent } from '@ionic/angular';
+import { toBoolean } from '@/core/transforms/boolean';
 import { Subscription } from 'rxjs';
-
 import { CoreSites } from '@services/sites';
-import { CoreFilepool } from '@services/filepool';
+import { CoreEditorService } from '@features/editor/services/editor';
+import { CoreEditorOffline } from '@features/editor/services/editor-offline';
+import { CoreEditorDraft } from '@features/editor/services/database/editor';
+import { NgZone } from '@singletons';
 import { CoreUrl } from '@singletons/url';
 import { CoreUtils } from '@singletons/utils';
-import { CoreEventFormActionData, CoreEventObserver, CoreEvents } from '@singletons/events';
-import { CoreEditorOffline } from '../../services/editor-offline';
+import { CoreFilepool } from '@services/filepool';
 import { CoreDirectivesRegistry } from '@singletons/directives-registry';
 import { CoreLoadingComponent } from '@components/loading/loading';
-import { CoreScreen } from '@services/screen';
-import { CoreCancellablePromise } from '@classes/cancellable-promise';
-import { CoreDom } from '@singletons/dom';
-import { CorePlatform } from '@services/platform';
-import { Swiper } from 'swiper';
-import { SwiperOptions } from 'swiper/types';
-import { ContextLevel } from '@/core/constants';
-import { CoreSwiper } from '@singletons/swiper';
-import { CoreWait } from '@singletons/wait';
-import { toBoolean } from '@/core/transforms/boolean';
-import { CoreQRScan } from '@services/qrscan';
-import { CoreSharedModule } from '@/core/shared.module';
-import { CoreSettingsHelper } from '@features/settings/services/settings-helper';
+import { CoreToasts } from '@services/overlays/toasts';
+import { CorePromiseUtils } from '@singletons/promise-utils';
+import { convertTextToHTMLElement } from '@/core/utils/create-html-element';
 
 /**
- * Component to display a rich text editor if enabled.
- *
- * If enabled, this component will show a rich text editor. Otherwise it'll show a regular textarea.
+ * Component that displays a rich text editor.
  *
  * Example:
  * <core-rich-text-editor [control]="control" [placeholder]="field.name"></core-rich-text-editor>
  */
 @Component({
     selector: 'core-rich-text-editor',
-    templateUrl: 'core-editor-rich-text-editor.html',
+    templateUrl: 'core-rich-text-editor.html',
     styleUrl: 'rich-text-editor.scss',
     standalone: true,
     imports: [
@@ -69,12 +66,10 @@ import { CoreSettingsHelper } from '@features/settings/services/settings-helper'
     ],
     schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
-export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit, OnDestroy {
+export class CoreEditorRichTextEditorComponent implements AfterViewInit, OnDestroy, OnInit {
 
-    // Based on: https://github.com/judgewest2000/Ionic3RichText/
-    // @todo Anchor button, fullscreen...
-    // @todo Textarea height is not being updated when editor is resized. Height is calculated if any css is changed.
-    // @todo Implement ControlValueAccessor https://angular.io/api/forms/ControlValueAccessor.
+    private static readonly MIN_HEIGHT = 200; // Minimum height of the editor.
+    private static readonly DRAFT_AUTOSAVE_FREQUENCY = 30000;
 
     @Input() placeholder = ''; // Placeholder to set in textarea.
     @Input() control?: FormControl<string | undefined | null>; // Form control.
@@ -86,806 +81,177 @@ export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit,
     @Input() contextInstanceId?: number; // The instance ID related to the context.
     @Input() elementId?: string; // An ID to set to the element.
     @Input() draftExtraParams?: Record<string, unknown>; // Extra params to identify the draft.
-    @Output() contentChanged: EventEmitter<string | undefined | null>;
+    @Output() contentChanged: EventEmitter<string | undefined | null> = new EventEmitter();
 
-    protected editorElement?: HTMLDivElement; // WYSIWYG editor.
-    @ViewChild('editor') editor?: ElementRef<HTMLDivElement>;
+    @ViewChild(CoreDynamicComponent) dynamicComponent!: CoreDynamicComponent<CoreEditorBaseComponent>;
 
-    @ViewChild('toolbar') toolbar?: ElementRef<HTMLDivElement>;
-
-    protected textareaElement?: HTMLTextAreaElement;
-    @ViewChild('textarea') textarea?: IonTextarea; // Textarea editor.
-
-    protected toolbarSlides?: Swiper;
-    @ViewChild('swiperRef') set swiperRef(swiperRef: ElementRef) {
-        /**
-         * This setTimeout waits for Ionic's async initialization to complete.
-         * Otherwise, an outdated swiper reference will be used.
-         */
-        setTimeout(async () => {
-            await this.waitLoadingsDone();
-
-            const swiper = CoreSwiper.initSwiperIfAvailable(this.toolbarSlides, swiperRef, this.swiperOpts);
-            if (!swiper) {
-                return;
-            }
-
-            this.toolbarSlides = swiper;
-        });
-    }
-
-    protected readonly DRAFT_AUTOSAVE_FREQUENCY = 30000;
-    protected readonly RESTORE_MESSAGE_CLEAR_TIME = 6000;
-    protected readonly SAVE_MESSAGE_CLEAR_TIME = 2000;
-
-    protected element: HTMLElement;
-    protected minHeight = 200; // Minimum height of the editor.
-
-    protected valueChangeSubscription?: Subscription;
     protected keyboardObserver?: CoreEventObserver;
-    protected resetObserver?: CoreEventObserver;
+    protected resizeListener?: CoreEventObserver;
+    protected editorComponentClass?: Type<CoreEditorBaseComponent>;
+    protected editorComponentData: Record<string, unknown> = {};
+    protected controlSubscription?: Subscription;
     protected labelObserver?: MutationObserver;
-    protected contentObserver?: MutationObserver;
-    protected initHeightInterval?: number;
-    protected isCurrentView = true;
-    protected toolbarButtonWidth = 44;
-    protected toolbarArrowWidth = 44;
+    protected setContentId = 0;
+
+    // Autosave.
     protected pageInstance: string;
-    protected autoSaveInterval?: number;
-    protected hideMessageTimeout?: number;
     protected lastDraft = '';
     protected draftWasRestored = false;
     protected originalContent?: string;
-    protected resizeFunction?: () => Promise<number>;
-    protected selectionChangeFunction = (): void => this.updateToolbarStyles();
-    protected resizeListener?: CoreEventObserver;
-    protected domPromise?: CoreCancellablePromise<void>;
-    protected buttonsDomPromise?: CoreCancellablePromise<void>;
-    protected shortcutCommands?: Record<string, EditorCommand>;
-    protected blurTimeout?: number;
-
-    rteEnabled = false;
-    isPhone = false;
-    toolbarHidden = false;
-    toolbarArrows = false;
-    toolbarPrevHidden = true;
-    toolbarNextHidden = false;
-    canScanQR = false;
-    ariaLabelledBy?: string;
-    infoMessage?: string;
-    toolbarStyles = {
-        strong: 'false',
-        b: 'false',
-        em: 'false',
-        i: 'false',
-        u: 'false',
-        strike: 'false',
-        p: 'false',
-        h3: 'false',
-        h4: 'false',
-        h5: 'false',
-        ul: 'false',
-        ol: 'false',
-    };
-
-    isEmpty = true;
-
-    swiperOpts: SwiperOptions = {
-        slidesPerView: 6,
-        centerInsufficientSlides: true,
-        watchSlidesProgress: true,
-    };
+    protected autoSaveInterval?: number;
+    protected resetObserver?: CoreEventObserver;
+    protected element: HTMLElement;
 
     constructor(
         @Optional() protected content: IonContent,
-        elementRef: ElementRef,
+        elementRef: ElementRef<HTMLElement>,
     ) {
-        this.contentChanged = new EventEmitter<string>();
+         // Generate a "unique" ID based on timestamp.
+        this.pageInstance = `app_${Date.now()}`;
         this.element = elementRef.nativeElement;
-        this.pageInstance = `app_${Date.now()}`; // Generate a "unique" ID based on timestamp.
-    }
-
-    /**
-     * @inheritdoc
-     */
-    ngOnInit(): void {
-        this.canScanQR = CoreQRScan.canScanQR();
-        this.isPhone = CoreScreen.isMobile;
-        this.toolbarHidden = this.isPhone;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    async ngAfterViewInit(): Promise<void> {
-        this.rteEnabled = await CoreSettingsHelper.isRichTextEditorEnabled();
-
-        await this.waitLoadingsDone();
-
-        // Setup the editor.
-        this.editorElement = this.editor?.nativeElement as HTMLDivElement;
-        this.textareaElement = await this.textarea?.getInputElement();
-        this.setContent(this.control?.value);
-        this.originalContent = this.control?.value ?? undefined;
-        this.lastDraft = this.control?.value ?? '';
-
-        // Use paragraph on enter.
-        // eslint-disable-next-line deprecation/deprecation
-        document.execCommand('DefaultParagraphSeparator', false, 'p');
-
-        this.maximizeEditorSize();
-
-        this.setListeners();
-        this.updateToolbarButtons();
-
-        if (this.elementId) {
-            // Prepend elementId with 'id_' like in web. Don't use a setter for this because the value shouldn't change.
-            this.elementId = `id_${this.elementId}`;
-            this.element.setAttribute('id', this.elementId);
-        }
-
-        // Update tags for a11y.
-        this.replaceTags(['b', 'i'], ['strong', 'em']);
-
-        if (this.shouldAutoSaveDrafts()) {
-            this.restoreDraft();
-
-            this.autoSaveDrafts();
-
-            this.deleteDraftOnSubmitOrCancel();
-        }
-
-        this.setupIonItem();
-
-        if (this.editorElement) {
-            const debounceMutation = CoreUtils.debounce(() => {
-                this.onChange();
-            }, 20);
-
-            this.contentObserver = new MutationObserver(debounceMutation);
-            this.contentObserver.observe(this.editorElement, { childList: true, subtree: true, characterData: true });
-        }
-    }
-
-    /**
-     * Setup Ion Item adding classes and managing aria-labelledby.
-     */
-    protected setupIonItem(): void {
-        const ionItem = this.element.closest<HTMLIonItemElement>('ion-item');
-        if (!ionItem) {
-            return;
-        }
-        ionItem.classList.add('item-rte');
-
-        const label = ionItem.querySelector('ion-label');
-        if (!label) {
-            return;
-        }
-
-        const updateArialabelledBy = () => {
-            this.ariaLabelledBy = label.getAttribute('id') ?? undefined;
-        };
-
-        this.labelObserver = new MutationObserver(updateArialabelledBy);
-        this.labelObserver.observe(label, { attributes: true, attributeFilter: ['id'] });
-
-        // Usually the label won't have an id, so we need to add one.
-        if (!label.getAttribute('id')) {
-            label.setAttribute('id', `rte-${CoreUtils.getUniqueId('CoreEditorRichTextEditor')}`);
-        }
-
-        updateArialabelledBy();
-    }
-
-    /**
-     * Set listeners and observers.
-     */
-    protected setListeners(): void {
-        // Listen for changes on the control to update the editor (if it is updated from outside of this component).
-        this.valueChangeSubscription = this.control?.valueChanges.subscribe((newValue) => {
-            if (this.draftWasRestored && this.originalContent === newValue) {
-                // A draft was restored and the content hasn't changed in the site. Use the draft value instead of this one.
-                this.control?.setValue(this.lastDraft, { emitEvent: false });
-
-                return;
-            }
-
-            // Apply the new content.
-            this.setContent(newValue);
-            this.originalContent = newValue ?? undefined;
-            this.infoMessage = undefined;
-
-            // Save a draft so the original content is saved.
-            this.lastDraft = newValue ?? '';
-            CoreEditorOffline.saveDraft(
-                this.contextLevel || ContextLevel.SYSTEM,
-                this.contextInstanceId || 0,
-                this.elementId || '',
-                this.draftExtraParams || {},
-                this.pageInstance,
-                this.lastDraft,
-                this.originalContent,
-            );
-        });
-
-        this.resizeListener = CoreDom.onWindowResize(() => {
-            this.windowResized();
-        }, 50);
-
-        document.addEventListener('selectionchange', this.selectionChangeFunction);
-
-        this.keyboardObserver = CoreEvents.on(CoreEvents.KEYBOARD_CHANGE, () => {
-            // Opening or closing the keyboard also calls the resize function, but sometimes the resize is called too soon.
-            // Check the height again, now the window height should have been updated.
-            this.maximizeEditorSize();
-        });
-    }
-
-    /**
-     * Handle keydown events in the editor.
-     *
-     * @param event Event
-     */
-    onKeyDown(event: KeyboardEvent): void {
-        const shortcutId = this.getShortcutId(event);
-        const commands = this.getShortcutCommands();
-        const command = commands[shortcutId];
-
-        if (!command) {
-            return;
-        }
-
-        this.stopBubble(event);
-        this.executeCommand(command);
-    }
-
-    /**
-     * Resize editor to maximize the space occupied.
-     */
-    protected async maximizeEditorSize(): Promise<void> {
-        // Editor is ready, adjust Height if needed.
-        const blankHeight = await this.getBlankHeightInContent();
-        const newHeight = blankHeight + this.element.getBoundingClientRect().height;
-
-        if (newHeight > this.minHeight) {
-            this.element.style.setProperty('--core-rte-height', `${newHeight - 1}px`);
-        } else {
-            this.element.style.removeProperty('--core-rte-height');
-        }
-    }
-
-    /**
-     * Wait until all <core-loading> children inside the page.
-     *
-     * @returns Promise resolved when loadings are done.
-     */
-    protected async waitLoadingsDone(): Promise<void> {
-        this.domPromise = CoreDom.waitToBeInDOM(this.element);
-
-        await this.domPromise;
-
-        const page = this.element.closest('.ion-page');
-        if (!page) {
-            return;
-        }
-
-        await CoreDirectivesRegistry.waitDirectivesReady(page, 'core-loading', CoreLoadingComponent);
-    }
-
-    /**
-     * Get the height of the space in blank at the end of the page.
-     *
-     * @returns Blank height in px. Will be negative if no blank space.
-     */
-    protected async getBlankHeightInContent(): Promise<number> {
-        await CoreWait.nextTicks(5); // Ensure content is completely loaded in the DOM.
-
-        let content: Element | null = this.element.closest('ion-content');
-        const contentHeight = await CoreDom.getContentHeight(this.content);
-
-        // Get first children with content, not fixed.
-        let scrollContentHeight = 0;
-        while (scrollContentHeight === 0 && content?.children) {
-            const children = Array.from(content.children)
-                .filter((element) => element.slot !== 'fixed' && !element.classList.contains('core-loading-container'));
-
-            scrollContentHeight = children
-                .map((element) => element.getBoundingClientRect().height)
-                .reduce((a,b) => a + b, 0);
-
-            content = children[0];
-        }
-
-        return contentHeight - scrollContentHeight;
-    }
-
-    /**
-     * On change function to sync with form data.
-     */
-    onChange(): void {
-        if (this.rteEnabled) {
-            if (!this.editorElement) {
-                return;
-            }
-
-            if (this.isNullOrWhiteSpace(this.editorElement)) {
-                this.clearText();
-            } else {
-                // The textarea and the form control must receive the original URLs.
-                this.restoreExternalContent();
-                // Don't emit event so our valueChanges doesn't get notified by this change.
-                this.control?.setValue(this.editorElement.innerHTML, { emitEvent: false });
-                this.control?.markAsDirty();
-                if (this.textarea) {
-                    this.textarea.value = this.editorElement.innerHTML;
-                }
-                // Treat URLs again for the editor.
-                this.treatExternalContent();
-            }
-        } else {
-            if (!this.textarea) {
-                return;
-            }
-
-            if (this.isNullOrWhiteSpace(this.textarea.value)) {
-                this.clearText();
-            } else {
-                // Don't emit event so our valueChanges doesn't get notified by this change.
-                this.control?.setValue(this.textarea.value, { emitEvent: false });
-                this.control?.markAsDirty();
-            }
-        }
-
-        this.contentChanged.emit(this.control?.value);
-    }
-
-    /**
-     * Set the caret position on the character number.
-     *
-     * @param parent Parent where to set the position.
-     */
-    protected setCurrentCursorPosition(parent: Node): void {
-        if (!this.rteEnabled || !this.element.classList.contains('has-focus')) {
-            return;
-        }
-
-        const range = document.createRange();
-
-        // Select all so it will go to the end.
-        range.selectNode(parent);
-        range.selectNodeContents(parent);
-        range.collapse(false);
-
-        const selection = window.getSelection();
-        if (!selection) {
-            return;
-        }
-        selection.removeAllRanges();
-        selection.addRange(range);
-    }
-
-    /**
-     * Toggle from rte editor to textarea syncing values.
-     *
-     * @param event The event.
-     */
-    async toggleEditor(event: Event): Promise<void> {
-        if (event.type === 'keyup' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
-            return;
-        }
-
-        this.stopBubble(event);
-
-        // Update tags for a11y.
-        this.replaceTags(['b', 'i'], ['strong', 'em']);
-
-        this.setContent(this.control?.value || '');
-
-        this.rteEnabled = !this.rteEnabled;
-
-        // Set focus and cursor at the end.
-        // Modify the DOM directly so the keyboard stays open.
-
-        if (this.rteEnabled) {
-            this.editorElement?.removeAttribute('hidden');
-            this.textareaElement?.setAttribute('hidden', '');
-        } else {
-            this.editorElement?.setAttribute('hidden', '');
-            this.textareaElement?.removeAttribute('hidden');
-        }
-
-        await CoreWait.nextTick();
-
-        this.focusRTE(event);
-    }
-
-    /**
-     * Treat elements that can contain external content.
-     * We only search for images because the editor should receive unfiltered text, so the multimedia filter won't be applied.
-     * Treating videos and audios in here is complex, so if a user manually adds one he won't be able to play it in the editor.
-     */
-    protected treatExternalContent(): void {
-        if (!CoreSites.isLoggedIn() || !this.editorElement) {
-            // Only treat external content if the user is logged in.
-            return;
-        }
-
-        const elements = Array.from(this.editorElement.querySelectorAll('img'));
-        const site = CoreSites.getCurrentSite();
-        const siteId = CoreSites.getCurrentSiteId();
-        const canDownloadFiles = !site || site.canDownloadFiles();
-        elements.forEach(async (el) => {
-            if (el.getAttribute('data-original-src')) {
-                // Already treated.
-                return;
-            }
-
-            const url = el.src;
-
-            if (!url || !CoreUrl.isDownloadableUrl(url) || (!canDownloadFiles && site?.isSitePluginFileUrl(url))) {
-                // Nothing to treat.
-                return;
-            }
-
-            // Check if it's downloaded.
-            const finalUrl = await CoreFilepool.getSrcByUrl(siteId, url, this.component, this.componentId);
-
-            // Check again if it's already treated, this function can be called concurrently more than once.
-            if (!el.getAttribute('data-original-src')) {
-                el.setAttribute('data-original-src', el.src);
-                el.setAttribute('src', finalUrl);
-            }
-        });
-    }
-
-    /**
-     * Reverts changes made by treatExternalContent.
-     */
-    protected restoreExternalContent(): void {
-        if (!this.editorElement) {
-            return;
-        }
-
-        const elements = Array.from(this.editorElement.querySelectorAll('img'));
-        elements.forEach((el) => {
-            const originalUrl = el.getAttribute('data-original-src');
-            if (originalUrl) {
-                el.setAttribute('src', originalUrl);
-                el.removeAttribute('data-original-src');
-            }
-        });
-    }
-
-    /**
-     * Check if text is empty.
-     *
-     * @param valueOrEl Text or element containing the text.
-     * @returns If value is null only a white space.
-     */
-    protected isNullOrWhiteSpace(valueOrEl: string | HTMLElement | null | undefined): boolean {
-        if (valueOrEl === null || valueOrEl === undefined) {
-            this.isEmpty = true;
-
-            return true;
-        }
-
-        this.isEmpty = typeof valueOrEl === 'string' ? CoreDom.htmlIsBlank(valueOrEl) : !CoreDom.elementHasContent(valueOrEl);
-
-        return this.isEmpty;
-    }
-
-    /**
-     * Set the content of the textarea and the editor element.
-     *
-     * @param value New content.
-     */
-    protected setContent(value: string | null | undefined): void {
-        if (!this.editorElement || !this.textarea) {
-            return;
-        }
-
-        if (this.isNullOrWhiteSpace(value)) {
-            // Avoid loops.
-            if (this.editorElement.innerHTML !== '<p></p>') {
-                this.editorElement.innerHTML = '<p></p>';
-            }
-            this.textarea.value = '';
-        } else {
-            value = value || '';
-            // Avoid loops.
-            if (this.editorElement.innerHTML !== value) {
-                this.editorElement.innerHTML = value;
-            }
-            this.textarea.value = value;
-            this.treatExternalContent();
-        }
     }
 
     /**
      * Clear the text.
      */
     clearText(): void {
-        this.setContent(null);
+        this.setContent('');
 
         // Don't emit event so our valueChanges doesn't get notified by this change.
         this.control?.setValue(null, { emitEvent: false });
-
-        setTimeout(() => {
-            if (this.editorElement) {
-                this.setCurrentCursorPosition(this.editorElement);
-            }
-        }, 1);
+        this.contentChanged.emit(null);
     }
 
     /**
-     * Execute an action over the selected text when a button is activated.
-     *  API docs: https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand
-     *
-     * @param event Event data
-     * @param command Command to execute.
-     * @param parameters If parameters is set to block, a formatBlock command will be performed. Otherwise it will switch the
-     *                      toolbar styles button when set.
+     * @inheritdoc
      */
-    buttonAction(event: Event, command: string, parameters?: string): void {
-        if (event.type === 'keyup' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
-            return;
-        }
-
-        this.stopBubble(event);
-
-        if (!command) {
-            return;
-        }
-
-        this.executeCommand({ name: command, parameters });
-    }
-
-    /**
-     * Execute an action over the selected text.
-     *  API docs: https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand
-     *
-     * @param command Editor command.
-     * @param command.name Command name.
-     * @param command.parameters Command parameters.
-     */
-    protected executeCommand({ name: command, parameters }: EditorCommand): void {
-        if (parameters === 'block') {
-            // eslint-disable-next-line deprecation/deprecation
-            document.execCommand('formatBlock', false, `<${command}>`);
-
-            return;
-        }
-
-        if (parameters) {
-            this.toolbarStyles[parameters] = this.toolbarStyles[parameters] == 'true' ? 'false' : 'true';
-        }
-
-        // eslint-disable-next-line deprecation/deprecation
-        document.execCommand(command, false);
-    }
-
-    /**
-     * Replace tags for a11y.
-     *
-     * @param originTags      Origin tags to be replaced.
-     * @param destinationTags Destination tags to replace.
-     */
-    protected replaceTags(originTags: string[], destinationTags: string[]): void {
-        if (!this.editorElement) {
-            return;
-        }
-
-        this.editorElement =
-            CoreDom.replaceTags(this.editorElement, originTags, destinationTags);
-
-        this.onChange();
-    }
-
-    /**
-     * Blur and hide the toolbar in phone mode.
-     *
-     * @param event Event.
-     */
-    blurRTE(event: FocusEvent): void {
-        const doBlur = (event: FocusEvent) => {
-            if (this.element.contains(document.activeElement)) {
-                // Do not hide if clicked inside the editor area, except hideButton.
-
-                return;
-            }
-
-            this.element.classList.remove('has-focus');
-
-            this.stopBubble(event);
-
-            if (this.isPhone) {
-                this.toolbarHidden = true;
-            }
+    async ngOnInit(): Promise<void> {
+        this.editorComponentClass = await CoreEditorService.getEditorComponentClass();
+        this.editorComponentData = {
+            name: this.name,
+            placeholder: this.placeholder,
+            component: this.component,
+            componentId: this.componentId,
+            contextLevel: this.contextLevel,
+            contextInstanceId: this.contextInstanceId,
+            onChange: this.onChange.bind(this),
         };
-
-        // There are many cases when focus is fired after blur, so we need to delay the blur action.
-        this.blurTimeout = window.setTimeout(() => doBlur(event),300);
     }
 
     /**
-     * Checks if Space or Enter have been pressed.
+     * Called when the editor content changes.
      *
-     * @param event Keyboard Event.
-     * @returns Wether space or enter have been pressed.
+     * @param content New content.
      */
-    protected isValidKeyboardKey(event: KeyboardEvent): boolean {
-        return event.key === ' ' || event.key === 'Enter';
-    }
-
-    /**
-     * Focus editor when click the area and show toolbar.
-     *
-     * @param event Event.
-     */
-    focusRTE(event: Event): void {
-        clearTimeout(this.blurTimeout);
-
-        if (this.rteEnabled) {
-            this.editorElement?.focus();
-        } else {
-            this.textarea?.setFocus();
-        }
-
-        this.element.classList.add('ion-touched');
-        this.element.classList.remove('ion-untouched');
-        this.element.classList.add('has-focus');
-
-        event && this.stopBubble(event);
-
-        this.toolbarHidden = false;
-        this.updateToolbarButtons();
-
-    }
-
-    /**
-     * Stop event default and propagation.
-     *
-     * @param event Event.
-     */
-    stopBubble(event: Event): void {
-        if (event.type != 'touchend' && event.type != 'mouseup' && event.type != 'keyup') {
-            event.preventDefault();
-        }
-        event.stopPropagation();
-    }
-
-    /**
-     * When a button is clicked first we should stop event propagation, but it has some cases to not.
-     *
-     * @param event Event.
-     */
-    downAction(event: Event): void {
-        if (event.type === 'keydown' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
-            return;
-        }
-
-        const selection = window.getSelection()?.toString();
-
-        // When RTE is focused with a whole paragraph in desktop the stopBubble will not fire click.
-        if (CorePlatform.isMobile() || !this.rteEnabled || document.activeElement != this.editorElement || selection === '') {
-            this.stopBubble(event);
-        }
-    }
-
-    /**
-     * Method that shows the next toolbar buttons.
-     */
-    async toolbarNext(event: Event): Promise<void> {
-        if (event.type === 'keyup' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
-            return;
-        }
-
-        this.stopBubble(event);
-
-        if (!this.toolbarNextHidden) {
-            const currentIndex = this.toolbarSlides?.activeIndex;
-            this.toolbarSlides?.slideTo((currentIndex || 0) + this.toolbarSlides.slidesPerViewDynamic());
-        }
-
-        await this.updateToolbarArrows();
-    }
-
-    /**
-     * Method that shows the previous toolbar buttons.
-     */
-    async toolbarPrev(event: Event): Promise<void> {
-        if (event.type === 'keyup' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
-            return;
-        }
-
-        this.stopBubble(event);
-
-        if (!this.toolbarPrevHidden) {
-            const currentIndex = this.toolbarSlides?.activeIndex;
-            this.toolbarSlides?.slideTo((currentIndex || 0) - this.toolbarSlides.slidesPerViewDynamic());
-        }
-
-        await this.updateToolbarArrows();
-    }
-
-    /**
-     * Update the number of toolbar buttons displayed.
-     */
-    async updateToolbarButtons(): Promise<void> {
-        if (!this.isCurrentView || !this.toolbar || !this.toolbarSlides ||
-            this.toolbarHidden || this.element.offsetParent === null) {
-            // Don't calculate if component isn't in current view, the calculations are wrong.
-            return;
-        }
-
-        const length = this.toolbarSlides.slides.length;
-
-        // Cancel previous one, if any.
-        this.buttonsDomPromise?.cancel();
-        this.buttonsDomPromise = CoreDom.waitToBeInDOM(this.toolbar?.nativeElement);
-        await this.buttonsDomPromise;
-
-        const width = this.toolbar?.nativeElement.getBoundingClientRect().width;
-
-        if (length > 0 && width > length * this.toolbarButtonWidth) {
-            this.swiperOpts.slidesPerView = length;
-            this.toolbarArrows = false;
-        } else {
-            this.swiperOpts.slidesPerView = Math.floor((width - this.toolbarArrowWidth * 2) / this.toolbarButtonWidth);
-            this.toolbarArrows = true;
-        }
-
-        await CoreWait.nextTick();
-
-        this.toolbarSlides.update();
-
-        await this.updateToolbarArrows();
-    }
-
-    /**
-     * Show or hide next/previous toolbar arrows.
-     */
-    async updateToolbarArrows(): Promise<void> {
-        if (!this.toolbarSlides) {
-            return;
-        }
-
-        const currentIndex = this.toolbarSlides.activeIndex;
-        const length = this.toolbarSlides.slides.length;
-        this.toolbarPrevHidden = currentIndex <= 0;
-        this.toolbarNextHidden = currentIndex + this.toolbarSlides.slidesPerViewDynamic() >= length;
-    }
-
-    /**
-     * Update highlighted toolbar styles.
-     */
-    updateToolbarStyles(): void {
-        const node = window.getSelection()?.focusNode;
-
-        if (!node || !this.element.contains(node)) {
-            return;
-        }
-
-        let element = node.nodeType === 1 ? node as HTMLElement : node.parentElement;
-
-        const styles = {};
-
-        while (element !== null && element !== this.editorElement) {
-            const tagName = element.tagName.toLowerCase();
-
-            if (this.toolbarStyles[tagName]) {
-                styles[tagName] = 'true';
+    protected onChange(content: string): void {
+        // Execute the callback in the Angular zone, so change detection doesn't stop working.
+        NgZone.run(() => {
+            if (CoreDom.htmlIsBlank(content)) {
+                this.control?.setValue(null, { emitEvent: false });
+                this.contentChanged.emit(null);
+            } else {
+                const container = document.createElement('template');
+                container.innerHTML = content;
+                this.restoreExternalContent(container.content);
+                content = container.innerHTML;
+                this.control?.setValue(content, { emitEvent: false });
+                this.control?.markAsDirty();
+                this.contentChanged.emit(content);
             }
-            element = element.parentElement;
+        });
+    }
+
+    /**
+     * @inheritdoc
+     */
+    async ngAfterViewInit(): Promise<void> {
+        if (this.elementId) {
+            // Prepend elementId with 'id_' like in web. Don't use a setter for this because the value shouldn't change.
+            this.elementId = 'id_' + this.elementId;
+            this.element.setAttribute('id', this.elementId);
         }
 
-        for (const tagName in this.toolbarStyles) {
-            this.toolbarStyles[tagName] = 'false';
+        this.maximizeEditorSize();
+
+        this.setupIonItem();
+
+        this.setContent(this.control?.value ?? '');
+
+        if (this.shouldAutoSaveDrafts()) {
+            this.restoreDraft();
+            this.autoSaveDrafts();
+            this.deleteDraftOnSubmitOrCancel();
         }
 
-        if (element === this.editorElement) {
-            Object.assign(this.toolbarStyles, styles);
+        this.resizeListener = CoreDom.onWindowResize(() => {
+            this.onWindowResize();
+        }, 50);
+
+        this.controlSubscription = this.control?.valueChanges.subscribe((newValue) => {
+            this.onControlValueChange(newValue);
+        });
+
+        // Opening or closing the keyboard also calls the resize function, but sometimes the resize is called too soon.
+        // Check the height again, now the window height should have been updated.
+        this.keyboardObserver = CoreEvents.on(CoreEvents.KEYBOARD_CHANGE, () => {
+            this.maximizeEditorSize();
+        });
+    }
+
+    /**
+     * @inheritdoc
+     */
+    ngOnDestroy(): void {
+        this.resizeListener?.off();
+        this.keyboardObserver?.off();
+        this.controlSubscription?.unsubscribe();
+        this.resetObserver?.off();
+        this.labelObserver?.disconnect();
+        clearInterval(this.autoSaveInterval);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    async onWindowResize(): Promise<void> {
+        await CoreWait.waitForResizeDone();
+
+        await this.maximizeEditorSize();
+
+        // Call onResize on editor implementation once it is loaded.
+        await this.dynamicComponent.ready();
+        await this.dynamicComponent.callComponentMethod('onResize');
+    }
+
+    /**
+     * Handles changes of the control value.
+     *
+     * @param newValue New value.
+     */
+    protected onControlValueChange(newValue: string | null | undefined): void {
+        if (this.draftWasRestored && this.originalContent === newValue) {
+            // A draft was restored and the content hasn't changed in the site. Use the draft value instead of this one.
+            this.control?.setValue(this.lastDraft, { emitEvent: false });
+            this.contentChanged.emit(this.lastDraft);
+
+            return;
         }
+
+        // Apply the new content.
+        this.setContent(newValue ?? '');
+        this.originalContent = newValue ?? undefined;
+
+        // Save a draft so the original content is saved.
+        this.lastDraft = newValue ?? '';
+        CoreEditorOffline.saveDraft(
+            this.contextLevel || ContextLevel.SYSTEM,
+            this.contextInstanceId || 0,
+            this.elementId || '',
+            this.draftExtraParams || {},
+            this.pageInstance,
+            this.lastDraft,
+            this.originalContent,
+        );
     }
 
     /**
@@ -907,8 +273,9 @@ export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit,
      * @returns Promise resolved when done.
      */
     protected async restoreDraft(): Promise<void> {
+        let entry: CoreEditorDraft | undefined;
         try {
-            const entry = await CoreEditorOffline.resumeDraft(
+            entry = await CoreEditorOffline.resumeDraft(
                 this.contextLevel || ContextLevel.SYSTEM,
                 this.contextInstanceId || 0,
                 this.elementId || '',
@@ -916,39 +283,40 @@ export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit,
                 this.pageInstance,
                 this.originalContent,
             );
-
-            if (entry === undefined) {
-                // No draft found.
-                return;
-            }
-
-            this.element.classList.add('ion-touched');
-            this.element.classList.remove('ion-untouched');
-
-            let draftText = entry.drafttext || '';
-
-            // Revert untouched editor contents to an empty string.
-            if (draftText === '<p></p>' || draftText === '<p><br></p>' || draftText === '<br>' ||
-                    draftText === '<p>&nbsp;</p>' || draftText === '<p><br>&nbsp;</p>') {
-                draftText = '';
-            }
-
-            if (draftText !== '' && this.control && draftText != this.control.value) {
-                // Restore the draft.
-                this.control.setValue(draftText, { emitEvent: false });
-                this.setContent(draftText);
-                this.lastDraft = draftText;
-                this.draftWasRestored = true;
-                this.originalContent = entry.originalcontent;
-
-                if (entry.drafttext != entry.originalcontent) {
-                    // Notify the user.
-                    this.showMessage('core.editor.textrecovered', this.RESTORE_MESSAGE_CLEAR_TIME);
-                }
-            }
         } catch {
             // Ignore errors, shouldn't happen.
         }
+
+        if (entry === undefined) {
+            // No draft found.
+            return;
+        }
+
+        let draftText = entry.drafttext ?? '';
+
+        // Revert untouched editor contents to an empty string.
+        if (CoreDom.htmlIsBlank(draftText)) {
+            draftText = '';
+        }
+
+        if (draftText !== '' && this.control && draftText != this.control.value) {
+            // Restore the draft.
+            this.setContent(draftText);
+            this.control.setValue(draftText, { emitEvent: false });
+            this.contentChanged.emit(draftText);
+            this.lastDraft = draftText;
+            this.draftWasRestored = true;
+            this.originalContent = entry.originalcontent;
+
+            if (entry.drafttext != entry.originalcontent) {
+                // Notify the user.
+                CoreToasts.show({
+                    message: 'core.editor.textrecovered',
+                    translateMessage: true,
+                });
+            }
+        }
+
     }
 
     /**
@@ -980,11 +348,14 @@ export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit,
 
                 // Draft saved, notify the user.
                 this.lastDraft = newText;
-                this.showMessage('core.editor.autosavesucceeded', this.SAVE_MESSAGE_CLEAR_TIME);
+                CoreToasts.show({
+                    message: 'core.editor.autosavesucceeded',
+                    translateMessage: true,
+                });
             } catch {
                 // Error saving draft.
             }
-        }, this.DRAFT_AUTOSAVE_FREQUENCY);
+        }, CoreEditorRichTextEditorComponent.DRAFT_AUTOSAVE_FREQUENCY);
     }
 
     /**
@@ -1010,210 +381,184 @@ export class CoreEditorRichTextEditorComponent implements OnInit, AfterViewInit,
     }
 
     /**
-     * Show a message.
-     *
-     * @param message Identifier of the message to display.
-     * @param timeout Number of milliseconds when to remove the message.
+     * Resize editor to maximize the space occupied.
      */
-    protected showMessage(message: string, timeout: number): void {
-        clearTimeout(this.hideMessageTimeout);
+    protected async maximizeEditorSize(): Promise<void> {
+        await this.waitLoadingsDone();
 
-        this.infoMessage = message;
+        // Editor is ready, adjust Height if needed.
+        const blankHeight = await this.getBlankHeightInContent();
+        const newHeight = blankHeight + this.element.getBoundingClientRect().height;
 
-        this.hideMessageTimeout = window.setTimeout(() => {
-            this.hideMessageTimeout = undefined;
-            this.infoMessage = undefined;
-        }, timeout);
+        if (newHeight > CoreEditorRichTextEditorComponent.MIN_HEIGHT) {
+            this.element.style.setProperty('--core-rte-height', `${newHeight - 1}px`);
+        } else {
+            this.element.style.removeProperty('--core-rte-height');
+        }
     }
 
     /**
-     * Scan a QR code and put its text in the editor.
+     * Get the height of the space in blank at the end of the page.
      *
-     * @param event Event data
-     * @returns Promise resolved when done.
+     * @returns Blank height in px. Will be negative if no blank space.
      */
-    async scanQR(event: Event): Promise<void> {
-        if (event.type === 'keyup' && !this.isValidKeyboardKey(<KeyboardEvent>event)) {
+    protected async getBlankHeightInContent(): Promise<number> {
+        await CoreDom.waitToBeInDOM(this.element);
+        await CoreWait.nextTicks(10);
+
+        let content: Element | null = this.element.closest('ion-content');
+        const contentHeight = await CoreDom.getContentHeight(this.content);
+
+        // Get first children with content, not fixed.
+        let scrollContentHeight = 0;
+        if (content) {
+            while (scrollContentHeight === 0 && content?.children) {
+                const children = Array.from(content.children)
+                    .filter((element) => element.slot !== 'fixed' && !element.classList.contains('core-loading-container'));
+
+                scrollContentHeight = children
+                    .map((element) => element.getBoundingClientRect().height)
+                    .reduce((a, b) => a + b, 0);
+
+                content = children[0];
+            }
+        }
+
+        return contentHeight - scrollContentHeight;
+    }
+
+    /**
+     * Set the content of the editor.
+     *
+     * @param content Content to set.
+     */
+    protected async setContent(content: string): Promise<void> {
+        const id = ++this.setContentId;
+        const container = convertTextToHTMLElement(content);
+
+        // Update tags for a11y.
+        CoreDom.replaceTags(container, ['b', 'i'], ['strong', 'em']);
+
+        try {
+            await this.treatExternalContent(container);
+        } finally {
+            // Set content on once the editor implementation is ready.
+            await this.dynamicComponent.ready();
+
+            // Only set the content if the function was not called again while treating external content.
+            if (id === this.setContentId) {
+                this.dynamicComponent.callComponentMethod('setContent', container.innerHTML);
+            }
+        }
+    }
+
+    /**
+     * Treat elements that can contain external content.
+     * We only search for images because the editor should receive unfiltered text, so the multimedia filter won't be applied.
+     * Treating videos and audios in here is complex, so if a user manually adds one he won't be able to play it in the editor.
+     *
+     * @param container Container with the content to treat.
+     */
+    protected async treatExternalContent(container: HTMLElement): Promise<void> {
+        if (!CoreSites.isLoggedIn()) {
+            // Only treat external content if the user is logged in.
             return;
         }
 
-        this.stopBubble(event);
+        const elements = Array.from(container.querySelectorAll('img'));
+        const site = CoreSites.getCurrentSite();
+        const siteId = CoreSites.getCurrentSiteId();
+        const canDownloadFiles = !site || site.canDownloadFiles();
+        const promises = elements.map(async (el) => {
+            if (el.getAttribute('data-original-src')) {
+                // Already treated.
+                return;
+            }
 
-        // Scan for a QR code.
-        const text = await CoreQRScan.scanQR();
+            const url = el.src;
 
-        if (text) {
-            this.focusRTE(event); // Make sure the editor is focused.
-            // eslint-disable-next-line deprecation/deprecation
-            document.execCommand('insertText', false, text);
-        }
+            if (!url || !CoreUrl.isDownloadableUrl(url) || (!canDownloadFiles && site?.isSitePluginFileUrl(url))) {
+                // Nothing to treat.
+                return;
+            }
+
+            // Check if it's downloaded.
+            const finalUrl = await CoreFilepool.getSrcByUrl(siteId, url, this.component, this.componentId);
+
+            // Check again if it's already treated, this function can be called concurrently more than once.
+            if (!el.getAttribute('data-original-src')) {
+                el.setAttribute('data-original-src', el.src);
+                el.setAttribute('src', finalUrl);
+            }
+        });
+
+        await CorePromiseUtils.allPromises(promises);
     }
 
     /**
-     * Window resized.
-     */
-    protected async windowResized(): Promise<void> {
-        await CoreWait.waitForResizeDone();
-        this.isPhone = CoreScreen.isMobile;
-
-        this.maximizeEditorSize();
-        this.updateToolbarButtons();
-    }
-
-    /**
-     * User entered the page that contains the component.
-     */
-    ionViewDidEnter(): void {
-        this.isCurrentView = true;
-
-        this.updateToolbarButtons();
-    }
-
-    /**
-     * User left the page that contains the component.
-     */
-    ionViewDidLeave(): void {
-        this.isCurrentView = false;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    ngOnDestroy(): void {
-        this.valueChangeSubscription?.unsubscribe();
-
-        document.removeEventListener('selectionchange', this.selectionChangeFunction);
-
-        clearInterval(this.initHeightInterval);
-        clearInterval(this.autoSaveInterval);
-        clearTimeout(this.hideMessageTimeout);
-
-        this.resetObserver?.off();
-        this.keyboardObserver?.off();
-        this.resizeListener?.off();
-
-        this.labelObserver?.disconnect();
-        this.contentObserver?.disconnect();
-
-        this.domPromise?.cancel();
-        this.buttonsDomPromise?.cancel();
-    }
-
-    /**
-     * Get commands triggered by keyboard shortcuts.
+     * Reverts changes made by treatExternalContent.
      *
-     * @returns Commands dictionary indexed by their corresponding keyboard shortcut id.
+     * @param container Container with the content to restore.
      */
-    getShortcutCommands(): Record<string, EditorCommand> {
-        if (!this.shortcutCommands) {
-            const isIOS = CorePlatform.isIOS();
-            const metaKey = isIOS ? 'metaKey' : 'ctrlKey';
-            const shiftKey = isIOS ? 'ctrlKey' : 'shiftKey';
-
-            // Same shortcuts as TinyMCE:
-            // @see https://www.tiny.cloud/docs/advanced/keyboard-shortcuts/
-            const shortcuts: { code: string; modifiers: (keyof KeyboardShortcut)[]; command: EditorCommand }[] = [
-                {
-                    code: 'KeyB',
-                    modifiers: [metaKey],
-                    command: {
-                        name: 'bold',
-                        parameters: 'strong',
-                    },
-                },
-                {
-                    code: 'KeyI',
-                    modifiers: [metaKey],
-                    command: {
-                        name: 'italic',
-                        parameters: 'em',
-                    },
-                },
-                {
-                    code: 'KeyU',
-                    modifiers: [metaKey],
-                    command: {
-                        name: 'underline',
-                        parameters: 'u',
-                    },
-                },
-                {
-                    code: 'Digit3',
-                    modifiers: ['altKey', shiftKey],
-                    command: {
-                        name: 'h3',
-                        parameters: 'block',
-                    },
-                },
-                {
-                    code: 'Digit4',
-                    modifiers: ['altKey', shiftKey],
-                    command: {
-                        name: 'h4',
-                        parameters: 'block',
-                    },
-                },
-                {
-                    code: 'Digit5',
-                    modifiers: ['altKey', shiftKey],
-                    command: {
-                        name: 'h5',
-                        parameters: 'block',
-                    },
-                },
-                {
-                    code: 'Digit7',
-                    modifiers: ['altKey', shiftKey],
-                    command: {
-                        name: 'p',
-                        parameters: 'block',
-                    },
-                },
-            ];
-
-            this.shortcutCommands = shortcuts.reduce((shortcuts, { code, modifiers, command }) => {
-                const id = this.getShortcutId({
-                    code: code,
-                    altKey: modifiers.includes('altKey'),
-                    metaKey: modifiers.includes('metaKey'),
-                    shiftKey: modifiers.includes('shiftKey'),
-                    ctrlKey: modifiers.includes('ctrlKey'),
-                });
-
-                shortcuts[id] = command;
-
-                return shortcuts;
-            }, {} as Record<string, EditorCommand>);
+    protected restoreExternalContent(container: DocumentFragment): void {
+        if (!CoreSites.isLoggedIn()) {
+            // Only treat external content if the user is logged in.
+            return;
         }
 
-        return this.shortcutCommands;
+        const elements = Array.from(container.querySelectorAll('img'));
+        elements.forEach((el) => {
+            const originalUrl = el.getAttribute('data-original-src');
+            if (originalUrl) {
+                el.setAttribute('src', originalUrl);
+                el.removeAttribute('data-original-src');
+            }
+        });
     }
 
     /**
-     * Get a unique identifier for a given keyboard shortcut.
+     * Wait until all <core-loading> children inside the page.
      *
-     * @param shortcut Shortcut.
-     * @returns Identifier.
+     * @returns Promise resolved when loadings are done.
      */
-    protected getShortcutId(shortcut: KeyboardShortcut): string {
-        return (shortcut.altKey ? '1' : '0')
-            + (shortcut.metaKey ? '1' : '0')
-            + (shortcut.shiftKey ? '1' : '0')
-            + (shortcut.ctrlKey ? '1' : '0')
-            + shortcut.code;
+    protected async waitLoadingsDone(): Promise<void> {
+        const page = this.element.closest('.ion-page');
+        if (!page) {
+            return;
+        }
+
+        await CoreDirectivesRegistry.waitDirectivesReady(page, 'core-loading', CoreLoadingComponent);
     }
 
-}
+    /**
+     * Setup Ion Item adding classes and managing aria-labelledby.
+     */
+    protected setupIonItem(): void {
+        const ionItem = this.element.closest<HTMLIonItemElement>('ion-item');
+        if (!ionItem) {
+            return;
+        }
+        ionItem.classList.add('item-rte');
 
-/**
- * Combination
- */
-type KeyboardShortcut = Pick<KeyboardEvent, 'code' | 'altKey' | 'metaKey' | 'ctrlKey' | 'shiftKey'>;
+        const label = ionItem.querySelector('ion-label');
+        if (!label) {
+            return;
+        }
 
-/**
- * Editor command.
- */
-interface EditorCommand {
-    name: string;
-    parameters?: string;
+        const updateArialabelledBy = async () => {
+            await this.dynamicComponent.ready();
+            this.editorComponentData.ariaLabelledBy = label.getAttribute('id') ?? undefined;
+        };
+
+        this.labelObserver = new MutationObserver(updateArialabelledBy);
+        this.labelObserver.observe(label, { attributes: true, attributeFilter: ['id'] });
+
+        // Usually the label won't have an id, so we need to add one.
+        if (!label.getAttribute('id')) {
+            label.setAttribute('id', 'rte-' + CoreUtils.getUniqueId('CoreEditorRichTextEditor'));
+        }
+
+        updateArialabelledBy();
+    }
+
 }
